requirements.txt
playwright==1.44.0
pytest==8.2.2
pytest-xdist==3.6.1
pytest-rerunfailures==13.0
allure-pytest==2.13.5
python-dotenv==1.0.1
